ml_pipeline_debugger/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-ml_pipeline_debugger"
version = "0.1.0"
description = "Ml Pipeline Debugger environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m ml_pipeline_debugger.server.app
server = "ml_pipeline_debugger.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["ml_pipeline_debugger", "ml_pipeline_debugger.server"]
package-dir = { "ml_pipeline_debugger" = ".", "ml_pipeline_debugger.server" = "server" }